Loading HTML pages with images in them no longer causes runtime crashes. :)
It was a pretty odd Flex event bug, exposed by one of the Flexlib containers. At the core of 
things, two different Flex events somehow end up bottoming out in the same magic string 
(I can't believe they're using non-unique strings for identifiers!), and the event dispatcher 
would die in very amusing ways, while trying to hand an event of one type to a handler 
expecting the other type. But now it's worked-around, and hopefully the next release of Flexlib 
will fix it as well.