For some reason (a bug in the new JDBC driver perhaps):
 floor(T1.rating)/?

is being seen as requiring two prepared statement arguments which is bizarre.

However, I don't see why we were dividing rating by two at all since we were
just using it as an order by expression. Removing the pointless division
"fixes" the problem, but we should beware of this potential bug in the future.