Added style sheet support for tutorial pages (and other HTML pages to be displayed inside flash).
The Flash usage of style sheets is slightly different from how web browsers
use them. In Flash, style sheets are not tied to the page being displayed (since
there is no "page" in the first place), but rather they're a property of the
HTML rendering control. Style sheet, once applied, persists across any changes
to the HTML being displayed. Correspondingly, Flash doesn't understand the
"<link rel=stylesheet>" tag for specifying an external CSS file inside HTML.

This code preserves the Flex stylesheet semantics, at the risk of confusing
users who expect the Flash HTML engine to behave like a web browser (which it
doesn't, in so many ways). In this code, a CSS file is loaded and applied
separately, by sending a new HELP_PAGE_SET_STYLE_COMMAND to the controller.
Once applied, the stylesheet persist across page loads, until replaced by
a different stylesheet.

For convenience, the function that opens a new tutorial page has been changed
to do both: load a page from a given URL, and also try to guess and apply
the appropriate CSS file. Stylesheet guessing is a trivial task of trying to load
a CSS file with the same root name as the web page (e.g. if the URL points
at some "foo.html", it will try to load the stylesheet "foo.css" as well).