And so it came to pass that DecorData was cast out into the fiery Gehenna, and its columns were dropped from the Scenes table, and its data scattered across the earth.
And only the Decor ID was kept in the scene, and used to load up all the decor info from the 
appropriate item table, because all the original data was getting pulled from there anyway, so 
it was getting cumbersome to duplicate it.

And this made the Decor Editor development easier, because now there weren't multiple copies 
of the same data laying around.

And there was much rejoicing.