Checkpoint commit for the "room editor v. 3". It's still a rough WIP. I'm pushing on getting the UI into a state where it can be evaluated, before I bring over any of the detailed functionality of the old editor. :)
There are some redundancies in this checkin, because I'm trying to keep the "v.2" editor 
in a working shape while also developing the "v.3" editor (hence a second edit icon for 
support+, double editing commands in the controller, etc). I'll be removing all of that 
scaffolding once v.3 is completely functional, of course.