If an avatar is updated (or deleted) while being worn, update the cached copy in the wearer's user object, update their occupant info.
This isn't actually quite right. What would be correct is to actually look
at the usage info for an avatar and update based on that. However, we don't
currently have a way to 'loan' items to other players, so this is
essentially the same. (Although, I believe it's possible to wear an avatar,
and then mail it to another, which is the same sort of thing.)
In other words: you should be able to say to a friend: "Here, wear this"
and offer them an avatar to wear without giving it to them. If they
wear something else it automatically takes it off, or if you go to wear
the same avatar it would see that it's used, look up that user and strip
them of it, before making it available for you.

So, when an avatar is updated, instead of merely checking to see if the owner
is wearing it, we should check the usage and update whoever is using it.
That would be TODO, because the usage info is undependable right now.
We need to fix that. Who's going to fix that?