Tip of the day, courtesy of MDB's School of Code Readability:
Avoid deep nesting by terminating control flow in a function rather than having
a function be an ever bifurcating set of if/then/else clauses.

To state this another way: get your sanity checking out of the way at the top
of the function and then do the real work in its unnested body.

For example, compare:

  protected function myDeeplyNestedAndHardToReadFunction () :void
  {
    if (somethingTotallyObvious) {
      if (someOtherExpectedThing) {
        // blah blah blah
        // blah blah blah
        if (omgWouldYouBelieveSomethingElse?) {
          // blah blah blah
          // blah blah blah
        } else {
          // zomg! badness
        }
      } else {
        // zomg! badness
      }
    } else {
      // zomg! badness
    }
  }

with:

  protected funtion myNotDeeplyNestedAndEasierToReadFunction () :void
  {
    if (!somethingTotallyObvious) {
      // zomg! badness
      return;
    }

    if (!someOtherExpectedThing) {
      // zomg! badness
      return;
    }

    // blah blah blah
    // blah blah blah

    if (!omgWouldYouBelieveSomethingElse?) {
      // zomg! badness
      return;
    }

    // blah blah blah
    // blah blah blah
  }

As one reads through the latter function, possible situations are introduced,
addressed, and allowed to gracefully depart one's short term memory by the
comforing "return" statement.

If/then/else constructs tax the brain by implying that one must remember the
entire "if" block so that when one gets to the "else" block, one can contrast
the different results of execution of the two blocks. Nesting conditionals
taxes the brain yet further as one is forced to deal with a geometrically
expandind set of possibilities.