Room editor updates:
Display a drop-down list of furni names in the room editor, letting the user
select which furni to edit without having to click on it.

The editor now queries the server for an entire array of names when it first
starts up, and each time a furni is added or removed from the room (but not
when it's merely being modified).