Grumble. Looks like I have to do a bunch of custom scene traversal befuckery for Whirled whether I like it or not because the originating scene knows where the client will arrive in the destination scene (which is a good thing) but if we have to switch servers on the way to the new scene, there's no way we're going to resolve the old scene, look up the old portal and then send the user to the new scene. So we need a new moveTo() that just passes the destination location directly. Which vastly simplifies everything and we can do away with our use of SceneService.moveTo() and SpotService.traversePortal() and always use the new method and everything will be lovely.
I contemplated putting this method into SpotService, but decided against it
because it's actually a lot simpler to have it in a sort of extension of
SceneService because the SceneDirector will need to keep the destination
location around while we're switching servers and reissue the moveTo() request
once we're on the target server, which turns out to be fiddly when it was the
SpotDirector that initiated the move via traversePortal(), but we can punt on
that for now since no one uses it. Scaling is so easy!