This check-in should fix the most glaring IE breakage, where IE silently fails to register Flash objects' external callback functions, making Flash completely inaccessible from Javascript (thus breaking the neighborhood view, adding furnis to rooms, wearing avatars, etc).
The breakage stemmed from what appears to be IE's implicit order-of-operation
assumption: if a Flash object is created at runtime as a DOM node, before
being added to the page, its callbacks will not be registered correctly. 
However, if an empty node is created first, added to the page, and only then
initialized with the Flash object's HTML definition, the foreign-function
interface will be registered correctly. 

Unfortunately, this bug constrains us in how we can create and initialize
Flash objects. Our code base often creates objects first, and then passes
their references around to be added to the appropriate containers. Because of
the bug, we need to co-locate object creation and placement as much as
possible. 

In some cases, it is enough to pass a definition string around, instead of a
full HTML widget, and then create the widget just before adding it to
the container. That's the approach used whenever possible.

But in some cases (e.g. avatar viewer) there are layers of indirection
between when the widget node (not necessarily Flash) is created, and when it
gets added to its container. Fortunately, in these cases, viewers don't need 
bi-directional communication, so I left them in their current shape, 
and changed the API and comments to make explicit what they don't do.

I have duplicated the work-around in two functions, WidgetUtil.addFlashObject(),
and Page.setFlashContent(). It currently seems impractical to try to unify them
further; Page objects have a particular way in which they initialize their
contents, which is different from generic containers.