Checkpoint commit for distributed chat channels.
After a couple of false starts, the winning solution uses channel DObjects
which get hosted by the one peer, and which are available for read-only
subscriptions to all other peers. Subscribers can't modify the chat channels 
directly - doing so led to data inconsistencies across the peer network, and 
issues with access control. Instead, a separate service now exists to allow
subscribers to ask the host to make changes - such as adding and removing
chatters from a channel. 

This is a WIP commit. Single-server chatting works as before, and much of the 
distributed functionality is there - one can join channels, chat, and leave 
channels, across different servers. But a couple of bugs remain to be ironed out, 
and I haven't yet addressed the issue of players switching servers while chatting.