SpeakService works very well for single-server chat, but distributed chat makes it tricky, because: 1) each peer has its own provider instance, so clients can't just refer to the one in the distributed object (that's the hosting peer's instance), and 2) only the hosting peer's provider should perform speak actions anyway.
For the distributed case, we give non-hosting peers a special provider, which forwards speak 
requests over to the hosting peer, instead of trying to perform them. But this means we need 
to tweak the distributed object, so that each client sees the provider instance on its peer 
server, and not the one on the host server. Yuck. 

This is still a WIP, because the Flash client needs some attention as well.