Implement a (possibly temporary) solution to a URL oddity.  Game lobbies are no longer bookmarkable - if you go to a game lobby from the flash client, nothing changes.  If you go from the Catalog or the MyWhirled view, the URL is changed to #game-NNN.  If you closed the game lobby, this still wouldn't change until you moved rooms, which means if you bookmarked the room, you'd actually be bookmarking the game.
Another odd effect was that if you browsed to the same game lobby from GWT twice in a row, it 
wouldn't work the second time because the GWT client thought we were just restoring our flash URL
and didn't send the parameters to the flash client.

Now the scene-id-style URL is restored immediately after the game lobby is opened, which fixes the 
issues above, but we still get a spurious game-related URL in the browser history, which doesn't 
really currently have a good solution.