Added a unique id to catalog listings separate from the item id of the listed item. Now that catalog listings are updated with updated items, this is especially more useful. This is also used now as the key for catalog listing comments, so that when the listed item is updated I don't have to renumber all the goddamned comments.
We should really consider moving item ratings and tags over to catalog listings
and nixing them on bare items because I still have to renumber those when a
catalog listing is updated and what does it really mean to tag or rate
something that's not listed in the catalog anyway? In any case, we can do that
later.