Add some actual tutorial events.
I'm realizing a problem with this approach, however; if the user goes to e.g. edit their profile without being in a scene (and subscribed to the Tutorial AVRG) then the Tutorial won't advance properly.

We may need to somehow 'store up' these events and poll them later. But I am making do with this for now, in the interest of momentum.