Fuck it. When a catalog listing is updated, everyone who owns that item is getting the updated version. We're bending over backwards to make games "magically update" and now I'd have to do a bunch of ballache to make content packs and trophies magically update, all to protect against some fuckwad turning his popular avatar into a penis. We're just going to cope with the penis.
It's also really nice for content developers to decide if they want to release
a whole new version of their product (a brand new avatar, etc.) or to just
magically update everyone who already has it for free. Let them deal with the
bitching if they change something that people liked.