I was the one whining about distclean. How am I supposed to know which targets there are? I can look through build.xml, but even that may not be enough because targets could be defined in an included file.
I realized there must be a command-line option to ant that displays targets.
There is, but it only shows the default and those with a description.
I've added some descriptions to some targets, so that running
ant -projecthelp can print out some sensible things. There may be more we
want to describe. Someone should do this for all our projects.