Checkpoint on the grand tab move.  Right now we get tabs in two places... but thats why this is a checkpoint.
So flexlib's SuperTabBar isn't really so Super.  As we've seen with SuperTabNavigator, its really
hard to get it to actually only allow some tabs to be closeable, and styling them properly has
also always been difficult.

Here I'm extending SuperTabBar, and am doing a bit of a hack to get the closeable-tabness correct.
I've also got it correctly showing the new tab as selected.  Finally, I should be able to minutely
control the way that the tabs look, as I have access to the actual tab instance (and can set the
style manually at the appropriate time, if necessary), so blinky red tabs for when we've got chat
waiting that we haven't seen yet should now be within the realm of possibility.

One last note:  We do get tabs in two locations, but the ones on the left have no bearing on the
chat that gets shown over the room - that comes next.