More steps towards the new chat order: ChannelChatTab is gone, replaced with ChatChannelController, which manages all the stuff the old one did, but isn't a container for chat itself.
This gets to the point where opening a chat channel does send back to the server and get the 
channel object, and closing the tab does close down the channel on the server.  You still can't
talk to the channel, or see the messages already in the channel though.  Nearing the end of 
checkpoint land.