Scratch that, reverse it. There were so many fucking problems with trying to do this that I would rather impale myself on a rusty pole than try to work around them. People can just load 10 fucking separate images for the navigation bar. Boy it sucks to develop for the goddamned web.
Case in point: Firefox puts 4 pixels of whitespace below an image and there's
not a thing you can do about it. Safari, being helpful, emulates this "feature"
so if you were like all of the other fuckwad web developers and decided that an
awesome "fix" would be to set "margin-top: -4px" on the thing below the image,
everything would look right. But of course IE isn't about to emulate one of
Firefox's bugs, so they put no such padding below their images.  I'll leave it
to the reader to imaging the twisty fucking maze of conditional CSS that would
be needed to do make things display properly.

Our previous solution was just to create a <div> with a hardcoded size and set
the background image to the image in question because that does not result in
any such padding on any browser (at least not any that I've tested with). But
we can't do that using GWT's fancy ImageBundle stuff. Performance--.