Another *almost-there* experiment with tabs, commented out.  I noticed that the tabs actually do a pretty good job of dealing with getting resized.  If the tab bar is too small, the tabs resize, and if a tab gets too small for the text it contains, it truncates the text intelligently, and gives the tab a tooltip with the full text.
I rolled this code out, which should stretch the spacer if the tabs are as wide as they want to be,
then once the spacer gets down to 0 width, shrink the tabs intead of cutting off the stuff to the
right of them.  And it works.  Except that sometimes for no reason I can find, the tabs just 
dissappear if you add or subtract a tab.  

I'm getting closer and closer to taking Ray's advice and just rolling my own tabs, extended straight
from UIComponent.  I'll make that decision when I revisit this, but first I want to muck about with
some of the stuff discussed in yesterday's meeting.