This just goes to show that no matter how much you try to avoid it, you can still get sucked down into flex hell.  Jon had some great swf symbols with scale-9 applied for the attention tab, and we've been trying for hours to get compiled world swf to recognize that scale 9 properly.
Now we've got something that works almost as well, but its requiring some wacky non-stylings that
were a pain to get working.

At any rate, we now have a shine effect that goes on when a tab contains text that you haven't seen
yet.  Also, the chat entry box changes color depending on the type of channel you're typing into.