Some days, I want to point the Fuck You hose at Sun instead of Adobe. For *eight fucking years* Sun has had bug reports and feature requests begging to provide a way to tell the goddamned AppletClassLoader not to fall back to trying to load resources directly from a URL if they're not found in the jar files specified in the applet tag.
News Flash Sun: as soon as JAR file support was added for applets (I believe
that might have been late 1996, maybe early 1997) everyone stopped downloading
their applets one classfile at a time. Everyone! Even the fucking kids in CS101
use a jar file! It would be super awesome if ClassLoader.findClass(),
ClassLoader.getResource() and ResourceBundle.getBundle() would stop fucking
making HTTP requests to try to find things that are not there. They will never
be there. Never! This stupid "feature" should fucking die die die!

Gee I wonder why applets are slow? Maybe it's because you're trying to fucking
reverse lookup the IP of the applet's HTTP server so that you can make sure
it's safe to make an HTTP request and ask it if just maybe they have a
goddamned Swahili translation of a properties file lying around in the document
root!

Anyway, miraculously we are able to work around this problem by creating a
URLClassLoader that omits everything except the jar files, explicitly giving it
no parent, carefully bing sure not to reference *any* non-system classes in our
top-level Applet and then passing everything through reflection into the actual
game which is loaded with our special classloader.

It still takes 5 seconds to resolve the class in the first place which is
probably Java deciding to redownload the Applet jar files or something
brilliant like that, but at least it's not the 90 seconds of network timeout
extravaganza that we had going on before. Fucking fuckers.