Conveniently, TextArea and other TextBoxBase derivations do not emit a ChangeEvent when you call setText() programmatically (it's probably a browser bug as they set the text in the underlying form element which should then report that it changed but it doesn't).
So we have to wrap TextArea and hijack setText(). We could stick with
CharacterLimiter and make a LimitedTextArea that extends TextArea, but it seems
more convenient to make a composite widget that manages its own label and can
be plugged right into wherever you would use a TextArea without a bunch of
fiddling and wiring up.