Slightly more evolved channel formats: Zell: Hallo. Zell ooO: I wish I had some icecream. Shout by Zell: I love ice cream! :: Zell loves ice cream. Broadcast by Zell: All players will be turned into soylent ice cream.
The idea is that instead of quotes we have a consistent colon followed by the message. I don't think either shouts or broadcasts actually get to the chat channels so we forgive them for being a bit convoluted. The think message is silly; I would actually suggest we only do Zell: for it, or simply forbid /think in channels.