Nothing like the smell of 5300 lines of diffs in the mornin'! Or evenin' as the case may be.
Major revamp to clean up the distinction between "base" (soon to be renamed
back to "msoy") and "world" so that we can factor the "game" stuff out
sufficiently that we can run games entirely in a separate client instance.

My Brawler memory leak investigations indicate that the problem goes much
deeper and that we (where we means Flash) leak not-insubstantial amounts of
memory every time we play a game even though we're loading everything in a
completely separate application domain and flushing the whole thing when we're
done, so it should not even be possible for a game to leak memory. Fortunately
Flash is so awesome that it is in fact possible.

Flash was clearly designed to run for a short period of time and then go away.
To the extent that we can allow it to do that, it is in our best interest to do
so. All of our diligent efforts not to leak memory will be dwarfed by the
vastly superior efforts of our sundry content creators who will take great
pains to write the least efficient, leakiest and worst possible code.

Let us not spend time polishing the wheelhouse whiel the whole boat sinks
slowly into the sea. Let us build a life raft instead.