More cleanups and wireups. It is now possible (with a hacked version of GWT) to load the game client, display the lobby and play a game, with nary a bit of world crap to be seen. Yay!
However, a bunch of stuff doesn't work in that case and all the crazy
world/game integration that we'll need in the "normal" case will take a bunch
more work. But still it's great to see this all running by its lonesome. It
fills me with hope.

Maybe I'll try Brawler in this mode and see how long it takes to crash. :)