Some very interesting CSS factoids:
1. Strict mode rendering is triggered by having an XML DOCTYPE, which we have.

2. In strict strict mode (rather than mostly strict mode, or quirks mode) img
is "display: inline" which means that it's like text and has blank space added
below it to make room for descenders (god knows why). Mostly strict mode
renders images as "display: block", we'll just specify that we want our images
rendered as block and voila, 2 pixel border is gone. Yay!

3. IE uses strict mode if a document has an XML DOCTYPE, except if it also has
an XML prologue, then it goes back to quirks mode. We had a prologue. We have
it no longer. This may fix a bunch of wacky IE problems and possibly allow us
to undo some of the fiddling Robert has had to do to get IE looking decent.

There are still baffling annoyances like a mysterious one pixel border below
the paged grid forward and next buttons on Linux Firefox (but not Mac or
Windows) and a one pixel border above and below *some* buttons in Safar (Edit
and Find Friends on the profile page) but not others that are in as far as I
can tell exactly the same constructs. I <heart> web browsers.