Track games and minutes played separately for single- and multiplayer play. This avoids payout problems when single-player games are much longer or shorter than multiplayer games.
I wrangled with trying to sensibly display both the single and multiplayer
average game durations and total game plays on the detail page and eventually
decided it was too confusing and too much information. So you see the total
game plays and the larger of the average durations.

Also the single player duration inherits the current stats because the only
instance where we know this is causing bad payouts is Underwhirled Drift who
will beneft from a clean slate in multiplayer durations.