The beginnings of the new all-signing all-dancing media chooser applet. Currently neither sings, nor dances.
What it does do is display an unsigned UI (a button that says "Change") which,
upon being clicked, loads a signed class, triggering the security dialog at
that time. Soon it will also display a pre-security-dialog warning to the user
that the security dialog may be coming and then store in a cookie that the user
has been duly warned so as to avoid repeating itself annoyingly in the future.