This was waiting for a release last week, and then there a couple minor bugs in Monday's release, so I thought there might be another...  It'll actually be relatively easy to disable this for a release if there is one before its ready for general consumption.
This is the channel occupant list.  Right now its really simple - just a list of horizontal white
boxes with the occupant's profile pic and name.  Working on scrolling now.  Also doesn't show room
occupants because it only pertains to Channels, which room chat is currently not.  Thats the next 
step - we want people to be able to subscribe to chat for a room that they don't currently occupy,
so room chat is going to have to act like a channel anyway (including cross-server hosting 
goodness).
 - I have come to the conclusion after this and chat tabs that my flex process needs to be thus:
   1) Am I doing *exactly* what the flex component was designed to do? Even including what the
      component creator thought should be stylable? Fine, use it.
   2) No? Don't even think about it.  Don't subclass it and attempt to force it into submission 
      though overridden methods - it wasn't designed to handle that.  Don't try to hack your 
      slightly different version from its super class - this is a headache too.  Just write your 
      own, starting with UIComponent or at the most, HBox/VBox.