- DataPackMediaContainers can load their media from bytes (although it currently   always assumes that the bytes represent a zip file). - The AvatarViewer now supports a "message" argument that is displayed if no media is   specified. - AvatarViewer can receive its media as a base64 encoded String through javascript...
As it turns out, that last one may be moot.

    Westheimer's Discovery: A couple of months in the laboratory can save a
    couple of hours in the library. -Frank H. Westheimer

God, I looked for a limit to the amount of data that can be sent between flash
and javascript, but I could not find it. So I had to experiment. As it turns out,
there is a limit around 1 megabyte. We're going to have to rethink the architecture of
remixing a bit, but I've already got a pretty good idea.