Make the background of the slide chat area white instead of transparent.  This uncovered a bug in the sizing of the slide chat container.  After spending way too long trying to figure out why it wasn't respecting the size I was giving it, I finally had an "oh shit!" moment, and realized what I'd been doing wrong for hours.
In digging, I also realized why (in slide mode), notification boxes in the scene were showing up in
the wrong place, and subsequently added a third chat overlay.  We now have one overlay for chat
that scroll with the scene (actor bubbles), one overlay for chat that doesn't move (notification
bubbles), and one overlay for the history list, which may move as a cohesive unit (when going into
slide mode).