Get this in before I have to merge anything more ;)
Couple of things here: We have a checkbox that shoes up to the left of the name in the current
room's chat channel. If you check it, you'll keep the channel when you move to a new room.  Right
now its disabled, because I'm waiting on newer smaller checkbox art from Jon.

Secondly, stop putting notifications in the chat history.  This was intially due to a bug that
mdb uncovered, but it makes sense anyway - notifications are temporary, and if you come back three
hours later to see that your friend logged on and off 15 times, its not really relevant.  Also,
now you'll stop staring at the new mail notification after you've already looked at it.  

Problems: The friend signing on/off notification isn't getting shown anywhere, because it never
been displayed in the upper right.  I'm going to change it next so it is, and then I'm going to
look into updating the notification dispatching so that it goes to the right tab, if there is one.
It can sit in that history, if you have it open.