We no longer need to be able to receive data via javascript.
Annoyingly, the avatarviewer compiles into a giant 1.1meg pile.
The reason for this is that actionscript is not Java, and we do things like
reference RoomController.AVATAR_CLICKED from inside the avatar sprite code,
and that links in the entire RoomController, and everything it references,
and so on. The avatarviewer ends up with the entire world client inside it.

I started to trace the dependancies to clean this up, but the problem
is epidemic and requires a massive cleanup- not a rabbit hole I'm going
down right this minute.