This precondition checking should really throw NullPointerException or IllegalArgumentException. As such we can just let the null pointer dereference trigger the NullPointerException and save ourselves the redundant code.
If we had annotations for expressing preconditions, I would advocate their use
as they don't sacrifice brevity and readability to express intention. The if
statements unfortunately distract from the actual code of the method.

The times where one should bite the bullet and introduce verbose precondition
checking are when the method in question *won't* trigger a NullPointerException
but rather allow one to happen later down the line where it will be confusing.
We want to throw NullPointerException as soon as possible. All of these methods
will already throw one as soon as possible.