Replaced all opaque "details" strings with structured data. Formatted strings were useful when dealing with text logs, but for our new logging system, it's more useful to retain transaction details in typed containers, so that we can store them as such later.
The details container is uniform for all transactions, and expects the player's member id and 
action type. Optionally, it also accepts information about the item that's the object of the 
transaction, and/or member id of any other player involved in it.

Also added more logging info for logins and game victories.