Added a background to HeaderBar. Moved window stylings into a window/ directory.
Of course the HeaderBar background doesn't work because Flex is an incorrigible
piece of shit. The header image, in spite of it being 20 pixels tall and the
HeaderBar definitely being 20 pixels tall (confirmed after all layout has
settled down), is being strangely scaled which breaks Rick's carefully
constructed pixels which are meant to align with the tab footer.

If I disabled backgroundSize: "100%", it renders the header background image
offset down by 2 pixels as if the header bar were 24 pixels tall, which as I
said before, it definitely is not. What the monkey fuck is going on I was not
able to discover in 45 minutes of fucking around which is about 44 minutes and
30 seconds longer than I should have had to spend on this code.