Apparently margin has no meaning on a <div>. If something inside a div has a margin, the browser gets to decide if that should mean the enclosing div gets bigger, or if the enclosing div remains the same size as its child that has a margin and (sort of) inherits it. Firefox decides to make the enclosing <div> bigger to accomodate the margins of its children, Safari (and IE, my bad Steve) choose to promote the margin to the enclosing <div>.
One might be inclined to suggest that one use padding on the inner element
instead of a margin. That would seem like a good idea but would actually just
get you punched. The inner element in this case is actually a table, and you
can't put padding around a table, IE inherits padding in its <td> from the
enclosing <table> element. So padding on a <table> completely breaks the
table. Thanks Bill (maybe I should blame Steve... Balmer that is).

HTML sucks. CSS sucks. We're nearly fifteen years along in the evolution of the
web and one still needs a PhD in Bullshitology to get anything to look half-way
decent in a goddamned web browser. I weep for the children.