Fix for the "clicking on game furni in IE doesn't do anything" bug. It actually did do something, too quickly to be visible: it opened the game panel and them immediately closed it down again.
Clicking on a game furni makes Flash posts a request to GWT to open the "#game" page. 
GWT obliges, opens the content pane, and minimizes the flash view. This in turn causes 
Flash to register itself for a click handler, so that clicking on it will restore it, and 
close the content pane.

In IE, all this happens too fast, and the "restore handler" registers for mouse clicks while 
the original furni click is still getting processed. As you probably guessed by now, the new 
restore handler would then get the original game furni click, thinks it's a separate click on 
the minimized flash window, and closed the lobby pane. Strangely enough, this only showed up 
in IE, and only when opening game panels.