Cleaned up a bunch of incorrect URL handling. The production system is a twisty maze of passages, so we have to be careful about how we do things.
The server_url contains the URL that should be used to talk to the production
server. This does not actually use the http_port configuration because the
servers listen on an HTTP port different from the one actually used to talk to
them due to the magic of our load balancing router and the PITA of getting a
Java process to listen on a privileged port.

So it's not safe to ever use http_port on the client, which is why we no longer
tell the client about http_port.