We used to clear out our world location when the game set its PlaceView which means that when (World)LocationDirector.locationDidChange(null) was dispatched, we would already "be" in the game.
Ray didn't like that we were in the world and game at the same time, so I
modified the liaison to leave our world location once we knew we were going to
enter the game (but were not already in it). That meant that
(World)LocationDirector.locationDidChange(null) was being dispatched when we
were nowhere. This was fine, but was further compounded by a bug I introduced
into the LocationDirector where it was not properly clearing out its place
information.

Now that bug is fixed, and the (Game)LocationDirector tells the WorldController
to update its location display once we've entered the game, so we can properly
update the HeaderBar and ControlBar to reflect our in-gameness.