I still can't reproduce it (after way too long spent attempting to), but I'm convinced this is the only place that we could be creating the "already added to channel" warning message on the server. The "is not in this channel" message still baffles me, but I'll take one hiesenbug at a time, tyvm.
There is probably something wonky going on in the client that would cause the channel handler to 
attempt to connect twice in a row, so when it does it's going to spew a stack trace, and if I ever
see it in my logs, I'll be able to look into it.  If you see it in your logs, please mail it on to
me.