Thread searching. This needs a little work:
1. We don't automatically apply your search terms to the matched thread when
you navigate to a thread that matches your search.

2. I'm not positive that the findThreads() query is not doing something stupid.
Anyone offhand know whether (warning, pseudo-SQL):

 select * from t1 join t2 on t1.threadId = t2.threadId
 where t1.groupId = G and (full_text_match(t1.subject, Q) or
                           full_text_match(t2.message, Q))

is going to execute full_text_match(t1.subject, Q) only once on the matched
thread records or if it's going to execute it repeatedly on the same thread
record for every message record that is paired with that thread by the join?

If it's not smart, I can just do two queries, one on the thread record subject
field and another on the message record message field and filter duplicates
myself by hand.