Fuck IE and fuck Firefox. Probably fuck Safari too, but I'm too annoyed to even check. IE is nice and buggy and we cannot put a position: absolute div in a position: relative div to overlap things unless any higher level element that is overflow: auto (or scroll) is also position: relative (or unless we're in quirks mode but we definitely don't want to go there). We used to only stuff our content into a scrolling div (read an overflow: auto div) if the browser was tall enough to display the client without scrolling. If the client had to scroll, then we said fuck it and let the whole page scroll. Well the whole page scrolling is apparently the "same" to the IE bug and that causes position: absolute in position: relative to break. So we *always* have to display our content in a position: relative container.
That means that we need to figure out how to scroll the client separately if
it's too tall to fit in the browser window. I'd just shrink the height of the
client but that violates our promise to games that they'll always get 500
pixels of vertical space and they'd break and players would become confused.
Instead we create a separate overflow: auto panel to contain the client. So if
the browser is too short the client gets its own scrollbar. Yay!

Next comes the fuck Firefox part (and/or Flash, let's spread the blame around).
On Firefox, if Flash is embedded in an overflow: auto (or scroll) element, for
no fucking explicable reason, the caret disappears from all text input fields.
Yay! So instead of giving 100% of our users a totally unusable text entry
experience, we only stuff the client into an overflow: auto element iff the
person's browser is too short to display the client. Yay!

Now if you'll excuse me I'm going to go jump off a bridge and save myself from
having to do any more web development.