Another reason I sometimes hate getters. There are a bunch of places in the flash libraries where getters/setters do special things and don't return an object that works the way you think it should. The classic example of this is the 'filters' propery on DisplayObjects, but the defaultTextFormat of a TextField appears to work the same.
It seems that when one accesses the defaultTextFormat property, you get a
clone of the TextFormat actually in use. Setting any new properties on that
without first saving it into a temporary variable is pointless, as you'll
be setting a property on the clone and then throwing the clone away.

So, when it seems that you could do
  field.defaultTextFormat.italic = true;
You actually need:
  var fmt :TextFormat = field.defaultTextFormat;
  fmt.italic = true;
  field.defaultTextFormat = fmt;